Consolidate tests and clean up
